Add zafClient
